+++ onnx/defs/quantization/defs.cc
@@ -104,7 +104,28 @@ ONNX_OPERATOR_SET_SCHEMA(
 
 static const char* DynamicQuantizeLinear_ver11_doc = R"DOC(
 A Function to fuse calculation for Scale, Zero Point and FP32->8Bit convertion of FP32 Input data.
-Outputs Scale, ZeroPoint and Quantized Input for a given FP32 Input)DOC";
+Outputs Scale, ZeroPoint and Quantized Input for a given FP32 Input.
+Scale is calculated as:
+```
+ y_scale = (max(x) - min(x))/(qmax - qmin)
+ * where qmax and qmin are max and min values for quantization range .i.e [0, 255] in case of uint8
+ * data range is adjusted to include 0.
+```
+Zero point is calculated as:
+```
+intermediate_zero_point = (qmin - min(x))/(qmax - qmin)
+y_zero_point = cast(round(saturate(itermediate_zero_point)))
+* where qmax and qmin are max and min values for quantization range .i.e [0, 255] in case of uint8
+* for saturation, it saturates to [0, 255] if it's uint8, or [-127, 127] if it's int8. Right now only uint8 is supported.
+* rounding to nearest ties to even.
+```
+Data quantization formula is:
+```
+y = saturate (round (x / y_scale) + y_zero_point)
+* for saturation, it saturates to [0, 255] if it's uint8, or [-127, 127] if it's int8. Right now only uint8 is supported.
+* rounding to nearest ties to even.
+```
+)DOC";
 
 ONNX_OPERATOR_SET_SCHEMA(
     DynamicQuantizeLinear,